feat: full cycle E2E test
